+++ examples/data-sources/meraki_organization_devices/data-source.tf
@@ -1,0 +1,3 @@
+data "meraki_organization_devices" "example" {
+  organization_id = "123456"
+}