+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - Add `meraki_organization_branding_policies_priorities` resource and data source
 - Add `meraki_network_client_splash_authorization_status` resource and data source
 - Add `meraki_network_devices` data source
+- Add `meraki_organization_devices` data source
 
 ## 0.1.2